print time

+++ cuda-workspace/TestingCudo/FastGridMain.c
@@ -163,18 +163,18 @@ void computeFull(StateList * initState,Path ** path, int N,int depth, int thread
 		double time_spent = (double) (end - begin) / CLOCKS_PER_SEC;
 		//printf("Time spent %lf\n", time_spent);
 		printf("Time spent %lf\n", time_spent);
-		double value = time_spent % 60;
-		printf("seconds %lf ", value);
+		int value = time_spent % 60;
+		printf("seconds %d ", value);
 		if(time_spent > 60)
 		{
 			time_spent /= 60;
 			value = time_spent%60;
-			printf("minutes %lf ", value);
+			printf("minutes %d ", value);
 			if(time_spent > 60)
 			{
 				time_spent /= 60;
 				value = time_spent%24;
-				printf("hours %lf ", vaule);
+				printf("hours %d ", vaule);
 				if(time_spent > 24)
 				{
 					time_spent /= 24;
